registration create & password crypt & some validation

+++ admin/includes/view_all_post.php
@@ -146,7 +146,7 @@
    echo "<td>$post_date</td>";
    echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
    echo "<td><a href='post.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-   echo "<td><a href='post.php?delete={$post_id}'>Delete</a></td>";
+   echo "<td><a onclick=\"javascript: return confirm('Are you sure you want to delete'); \" href='post.php?delete={$post_id}'>Delete</a></td>";
    echo "</tr>";
 }
